Add pax manifest page
