+++ models/boac/boac_departments.rb
@@ -1,27 +1,28 @@
 class BOACDepartments
 
-  attr_accessor :code, :name, :notes_only
+  attr_accessor :code, :name, :export_name, :notes_only
 
-  def initialize(code, name, notes_only=nil)
+  def initialize(code, name, export_name, notes_only=nil)
     @code = code
     @name = name
+    @export_name = export_name
     @notes_only = notes_only
   end
 
   DEPARTMENTS = [
-      ADMIN = new('ADMIN', 'Admins'),
-      ASC = new('UWASC', 'Athletic Study Center'),
-      CHEM = new('CDCDN', 'College of Chemistry'),
-      COE = new('COENG', 'College of Engineering'),
-      ENV_DESIGN = new('DACED', 'College of Environmental Design'),
-      GUEST = new('GUEST', 'Guest Access'),
-      HAAS = new('BAHSB', 'Haas School of Business'),
-      L_AND_S = new('QCADV', 'College of Letters & Science'),
-      L_AND_S_MAJ = new('QCADVMAJ', 'Letters & Science Major Advisors'),
-      NAT_RES = new('MANRD', 'College of Natural Resources'),
-      NOTES_ONLY = new('NOTESONLY', 'Notes Only', true),
-      OTHER = new('ZZZZZ', 'Other'),
-      ZCEEE = new('ZCEEE', 'Centers for Educational Equity and Excellence')
+      ADMIN = new('ADMIN', 'Admins', nil),
+      ASC = new('UWASC', 'Athletic Study Center', nil),
+      CHEM = new('CDCDN', 'College of Chemistry', nil),
+      COE = new('COENG', 'College of Engineering', nil),
+      ENV_DESIGN = new('DACED', 'College of Environmental Design', nil),
+      GUEST = new('GUEST', 'Guest Access', nil),
+      HAAS = new('BAHSB', 'Haas School of Business', nil),
+      L_AND_S = new('QCADV', 'College of Letters & Science', 'L&S College Advising'),
+      L_AND_S_MAJ = new('QCADVMAJ', 'Letters & Science Major Advisors', 'L&S Major Advising'),
+      NAT_RES = new('MANRD', 'College of Natural Resources', nil),
+      NOTES_ONLY = new('NOTESONLY', 'Notes Only', nil, true),
+      OTHER = new('ZZZZZ', 'Other', nil),
+      ZCEEE = new('ZCEEE', 'Centers for Educational Equity and Excellence', nil)
   ]
 
 end
