+++ util/spec_helper.rb
@@ -84,7 +84,6 @@ require_relative '../pages/boac/boac_group_modal_pages'
 require_relative '../pages/boac/boac_add_group_selector_pages'
 require_relative '../pages/boac/boac_list_view_pages'
 require_relative '../pages/boac/boac_user_list_pages'
-require_relative '../pages/boac/boac_admin_page'
 require_relative '../pages/boac/boac_home_page'
 require_relative '../pages/boac/boac_cohort_pages'
 require_relative '../pages/boac/boac_class_pages'
@@ -94,6 +93,8 @@ require_relative '../pages/boac/boac_group_page'
 require_relative '../pages/boac/boac_filtered_cohort_page_filters'
 require_relative '../pages/boac/boac_filtered_cohort_page_results'
 require_relative '../pages/boac/boac_filtered_cohort_page'
+require_relative '../pages/boac/boac_flight_deck_page'
+require_relative '../pages/boac/boac_pax_manifest_page'
 require_relative '../pages/boac/boac_search_results_page'
 require_relative '../pages/boac/boac_student_page_advising_note'
 require_relative '../pages/boac/boac_student_page'
